LPS-98111 Update to liferay-npm-scripts v4.8.0
Corresponds to public "master" PR:

https://github.com/brianchandotcom/liferay-portal/pull/75590

The big change there is that we now support cross-project imports in
frontend-tests, so I wanted to be sure that the tests on
"master-private" still pass.

Note that the test snapshots were stale, presumably because they should
have been updated in 73edbe98ecdcd58 but were not. I updated them, and
also deduplicated the yarn.lock.