LPS-99516 Remove jsdoc tooling
This was added in 747f244b07f6 ("LPS-91137 Create initial
documentation") back in February 2019, but as far as I can tell was
always broken because it relied on a plugin
(jsdoc-export-default-interop) that was never explicitly included in the
package.json file, nor transitively included via any other dependency.
On top of that, actually trying to run the jsdoc task would blow up
with:

    ERROR: Unable to find the source file or directory
    $ROOT/modules/private/apps/portal-workflow/portal-workflow-metrics-web/README.markdown

because that file was deleted in ae556a3047d9 ("LPS-91137 we don't have
this for any other *-web"), also from Febrary.

As I said in the previous commit, if nobody has noticed this breakage in
the intervening months, it seems appropriate to just remove it.

So this commit reverts what's left of the stuff added in 747f244b07f6,
including a "jsconfig.json" file which is an editor configuration file
and probably shouldn't be committed (we have no other files like this in
this repo or in liferay-portal).