LPS-101856 portal-search-tuning-synonyms-web: Fix multiple delete for entries with semicolon
When building the input value for rows to delete, use a comma and space (, ) as a delimiter instead of semicolon (;) since the multiselect input separates commas by default. Semicolons are allowed in a single synonym which can cause the synonyms to incorrectly divide synonym sets when attempting to delete.

Using only a comma as a delimiter will also not work because the value of a synonym set is separated using a comma already.